% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-icon.R
\name{progress_bar_icon}
\alias{progress_bar_icon}
\title{Progress icon type}
\usage{
progress_bar_icon(icon, length, width = options()$width - 20)
}
\arguments{
\item{icon}{The icon type you wish to use. See details}

\item{length}{The total number of iterations in the loop.}

\item{width}{Width of the bar. Default is width of console - 20 to allow room for other counters.}
}
\description{
Adds fun icons to your progress bar. A watched pot neer boils, ergo progress bars can feel like they take forever. May as well have some fun with it.
}
\details{
Icon types: \code{fish}, \code{tiefighter}, \code{dagger}, \code{greatsword}
}
\seealso{
[progress::progress_bar]
}
