% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-builder.R
\name{token}
\alias{token}
\alias{token.fish}
\alias{token.dagger}
\alias{token.tiefighter}
\title{Token builder}
\usage{
token(pbicon, j)

\method{token}{fish}(pbicon, j)

\method{token}{dagger}(pbicon, j)

\method{token}{tiefighter}(pbicon, j)
}
\arguments{
\item{pbicon}{Progress bar icon object}

\item{j}{The current iteration}
}
\description{
The token function build the string to pass to the progress bar object.
}
\details{
Each small fish represents 25\% of the progress. When they're all gobbled up by the big fish the job is done.

Each dagger box represents 20\% completion.

Each tie fighter represents 30\% completion.
}
\examples{
\dontrun{
 n <- 500
 bar_fmt <- ":elapsedfull | :fish |"
 pb <- progress_bar$new(format = bar_fmt, total = n, clear = FALSE)
 fish <- progress_bar_icon("fish", n)
 for(j in 1:n){
   pb$tick(tokens = list(
     fish = token(fish, j)
   ))
   Sys.sleep(0.03)
 }

 n <- 500
 bar_fmt <- green$bold(":elapsedfull | :tie |")
 pb <- progress_bar$new(format = bar_fmt, total = n, clear = FALSE)
 tie <- progress_bar_icon("tiefighter", n, 60)
 for(j in 1:n){
   pb$tick(tokens = list(
     tie = token(tie, j)
   ))
   Sys.sleep(0.03)
 }
}
}
