% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-icon.R
\name{progress_bar_icon}
\alias{progress_bar_icon}
\title{Progress icon type}
\usage{
progress_bar_icon(icon, length, width = options()$width - 20)
}
\arguments{
\item{icon}{The icon type you wish to use. See details}

\item{length}{The total number of iterations in the loop.}

\item{width}{Width of the bar. Default is width of console - 20 to allow room for other counters.}
}
\description{
Adds fun icons to your progress bar. A watched pot neer boils, ergo progress bars can feel like they take forever. May as well have some fun with it.
}
\details{
Icon types: \code{fish}, \code{tiefighter}, \code{dagger}, \code{greatsword}
}
\examples{
\dontrun{
 n <- 500
 bar_fmt <- ":elapsedfull | :fish |"
 pb <- progress_bar$new(format = bar_fmt, total = n, clear = FALSE)
 fish <- progress_bar_icon("fish", n)
 for(j in 1:n){
   pb$tick(tokens = list(
     fish = token(fish, j)
   ))
   Sys.sleep(0.03)
 }

 n <- 500
 bar_fmt <- green$bold(":elapsedfull | :tie |")
 pb <- progress_bar$new(format = bar_fmt, total = n, clear = FALSE)
 tie <- progress_bar_icon("tiefighter", n, 60)
 for(j in 1:n){
   pb$tick(tokens = list(
     tie = token(tie, j)
   ))
   Sys.sleep(0.03)
 }
}

}
\seealso{
[progress::progress_bar]
}
